+++ lib/Constants/FrontendMode.php
@@ -3,23 +3,18 @@
 namespace Heidelpay\PhpPaymentApi\Constants;
 
 /**
- * Class for PHP Payment API config & information constants
+ * Class for mode property contstants in frontend group.
  *
  * @license Use of this software requires acceptance of the License Agreement. See LICENSE file.
  * @copyright Copyright © 2017-present heidelpay GmbH. All rights reserved.
  *
  * @link http://dev.heidelpay.com/php-payment-api
  *
- * @author Stephano Vogel <development@heidelpay.com>
+ * @author Simon Gabriel <development@heidelpay.com>
  *
  * @package heidelpay\php-payment-api\constants
- *
- * @since 1.3.0 First time this was introduced.
  */
-class ApiConfig
+class FrontendMode
 {
-    const SDK_VERSION = 'v1.6.0';
-
-    const LIVE_URL = 'https://heidelpay.hpcgw.net/ngw/post';
-    const TEST_URL = 'https://test-heidelpay.hpcgw.net/ngw/post';
+    const FRONTEND_MODE_WHITELABEL = 'WHITELABEL';
 }