+++ tests/unit/ResponseTest.php
@@ -11,6 +11,7 @@ use Heidelpay\PhpPaymentApi\Exceptions\PaymentFormUrlException;
 use Heidelpay\PhpPaymentApi\Exceptions\HashVerificationException;
 use Heidelpay\PhpPaymentApi\Constants\TransactionMode;
 use Heidelpay\PhpPaymentApi\Constants\ProcessingResult;
+use Heidelpay\PhpPaymentApi\Constants\FrontendMode;
 
 /**
  *
@@ -86,7 +87,7 @@ class ResponseTest extends Test
             'IDENTIFICATION_SHOPPERID' => '12344',
             'PROCESSING_RETURN_CODE' => '000.100.112',
             'PROCESSING_RESULT' => ProcessingResult::ACK,
-            'FRONTEND_MODE' => 'WHITELABEL',
+            'FRONTEND_MODE' => FrontendMode::FRONTEND_MODE_WHITELABEL,
             'IDENTIFICATION_UNIQUEID' => '31HA07BC8108A9126F199F2784552637',
             'CRITERION_SECRET' => '209022666cd4706e5f451067592b6be1aff4a913d5bb7f8249f7418ee25c91b3' .
                 '18ebac66f41a6692539c8923adfdad6aae26138b1b3a7e37a197ab952be57876',
@@ -104,14 +105,16 @@ class ResponseTest extends Test
             'FRONTEND_LANGUAGE' => 'DE',
             'PAYMENT_CODE' => PaymentMethod::CREDIT_CARD . '.' . TransactionType::REGISTRATION,
             'BASKET_ID' => '31HA07BC8129FBB819367B2205CD6FB4',
-            'RISKINFORMATION_SINCE' => '2017-01-01',
-            'RISKINFORMATION_ORDERCOUNT' => '5',
-            'RISKINFORMATION_GUESTCHECKOUT' => 'FALSE',
+            'RISKINFORMATION_CUSTOMERSINCE' => '2017-01-01',
+            'RISKINFORMATION_CUSTOMERORDERCOUNT' => '5',
+            'RISKINFORMATION_CUSTOMERGUESTCHECKOUT' => 'FALSE',
             'CONNECTOR_ACCOUNT_HOLDER' => 'Test Account Holder',
             'CRITERION_TEST_VALUE' => 'Test Value',
             'INVALID_PROP' => 'Invalid',
             'lang' => 'de'
         );
+
+        ksort($this->responseArray);
 
         $this->responseObject = new Response($this->responseArray);
     }
@@ -443,5 +446,85 @@ class ResponseTest extends Test
     {
         $reponse = Response::fromPost([]);
         $this->assertEquals(Response::class, get_class($reponse));
+    }
+
+    /**
+     * Assert the object is being updated with the given rawResponseParameters and then returned.
+     *
+     * @test
+     */
+    public function splitArrayShouldUpdateAndReturnTheResponseObjectUsingTheGivenResponseArray()
+    {
+        $this->compareResponseArrayToActualArray($this->responseArray);
+
+        // change all parameters of the response
+        $manipulatedResponseArray = $this->getManipulatedArray();
+        $this->responseObject->splitArray($manipulatedResponseArray);
+
+        $this->compareResponseArrayToActualArray($manipulatedResponseArray);
+    }
+
+    /**
+     * Return a response array with manipulated values.
+     *
+     * @return array
+     */
+    private function getManipulatedArray()
+    {
+        $manipulatedResponseArray = [];
+
+        foreach ($this->responseArray as $key => $item) {
+            switch (true) {
+                case is_float($item):
+                    $item = 123.45;
+                    break;
+                case is_bool($item):
+                    $item = !$item;
+                    break;
+                case is_string($item):
+                    if ($key !== 'FRONTEND_MODE') {
+                        $item .= '--' . $key;
+                    }
+                    break;
+                default:
+                    // do nothing
+                    break;
+            }
+
+            $manipulatedResponseArray[$key] = $item;
+        }
+
+        return $manipulatedResponseArray;
+    }
+
+    /**
+     * Replaces the underscore of the raw response with a period to make it comparable.
+     *
+     * @return array
+     */
+    private function getComparableArray(array $responseArray = [])
+    {
+        $comparableResponseArray = [];
+        foreach ($responseArray as $key => $item) {
+            $comparableResponseArray[preg_replace('/_/', '.', $key, 1)] = $item;
+        }
+        return $comparableResponseArray;
+    }
+
+    /**
+     * Compares a given rawResponse with the properties of the response object by converting it back into an array.
+     *
+     * @param $responseArray
+     */
+    private function compareResponseArrayToActualArray($responseArray)
+    {
+        $comparableResponseArray = $this->getComparableArray($responseArray);
+        $responseObjectArray = $this->responseObject->toArray(true);
+
+        foreach ($comparableResponseArray as $key => $item) {
+            if (array_key_exists($key, $responseObjectArray)) {
+                $this->assertEquals($comparableResponseArray[$key], $responseObjectArray[$key]);
+            }
+        }
     }
 }